run_make_support: move `ar` into own module
